# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Estados",
        "description": "API para gestión de estados",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://uptyab.test/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener lista de usuarios",
                "operationId": "e9d4d8b419dbef04b5748fbf9ccf3881",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "4e9dbcbdaaaf9faab196a82d5cbaea23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rol_id",
                                    "cedula",
                                    "nombres",
                                    "apellidos",
                                    "email",
                                    "clave"
                                ],
                                "properties": {
                                    "rol_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cedula": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "V12345678"
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Juan"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 100,
                                        "example": "juan@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+584123456789",
                                        "nullable": true
                                    },
                                    "clave": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "password123"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener un usuario por ID",
                "operationId": "e9ac05ddc4ea46d8848b012d7d92e3ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario existente",
                "operationId": "27ac608b6dd8b8e7c7c45b10c5210b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rol_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cedula": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "V12345678"
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Juan"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 100,
                                        "example": "juan@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+584123456789",
                                        "nullable": true
                                    },
                                    "clave": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "nuevapassword"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "ea700764859a665afd38bb39c2842f1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Usuario": {
                "title": "Usuario",
                "description": "Modelo que representa un usuario del sistema",
                "properties": {
                    "id": {
                        "description": "ID del usuario",
                        "type": "integer",
                        "format": "int64"
                    },
                    "rol_id": {
                        "description": "ID del rol del usuario",
                        "type": "integer",
                        "format": "int64"
                    },
                    "cedula": {
                        "description": "Número de cédula del usuario",
                        "type": "string",
                        "maxLength": 20
                    },
                    "nombres": {
                        "description": "Nombres del usuario",
                        "type": "string",
                        "maxLength": 100
                    },
                    "apellidos": {
                        "description": "Apellidos del usuario",
                        "type": "string",
                        "maxLength": 100
                    },
                    "email": {
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "format": "email",
                        "maxLength": 100
                    },
                    "telefono": {
                        "description": "Número de teléfono del usuario",
                        "type": "string",
                        "maxLength": 20,
                        "nullable": true
                    },
                    "estado": {
                        "description": "Estado del usuario (activo/inactivo)",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}